move slice and string helpers into separate module
