+++ src/sqlfluff/core/helpers/__init__.py
@@ -1,0 +1,9 @@
+"""Helper methods for other classes.
+
+This module should have no dependencies on other parts of
+the SQLFluff package and each should also be free of dependencies
+on each other. This is to ensure each of these methods can be used
+anywhere within the project without fear of dependency issues.
+
+Methods are organised by the datatype they are designed to help with.
+"""
