+++ src/sqlfluff/core/helpers/slice.py
