+++ src/sqlfluff/core/helpers/string.py
