+++ src/sqlfluff/core/parser/grammar/base.py
@@ -15,13 +15,13 @@ from typing import (
 )
 from uuid import UUID, uuid4
 
+from sqlfluff.core.helpers.string import curtail_string
 from sqlfluff.core.parser.context import ParseContext
 from sqlfluff.core.parser.match_algorithms import greedy_match
 from sqlfluff.core.parser.match_result import MatchResult
 from sqlfluff.core.parser.matchable import Matchable
 from sqlfluff.core.parser.segments import BaseSegment
 from sqlfluff.core.parser.types import ParseMode, SimpleHintType
-from sqlfluff.core.string_helpers import curtail_string
 
 if TYPE_CHECKING:  # pragma: no cover
     from sqlfluff.core.dialects.base import Dialect
